+++ mne/io/base.py
@@ -40,7 +40,7 @@ from ..viz import plot_raw, plot_raw_psd, plot_raw_psd_topo
 from ..defaults import _handle_default
 from ..externals.six import string_types
 from ..event import find_events, concatenate_events
-from ..annotations import _combine_annotations, _onset_to_seconds
+from ..annotations import Annotations, _combine_annotations, _onset_to_seconds
 
 
 class ToDataFrameMixin(object):
@@ -584,6 +584,62 @@ class _BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
     @property
     def _raw_lengths(self):
         return [l - f + 1 for f, l in zip(self._first_samps, self._last_samps)]
+
+    @property
+    def annotations(self):
+        """Annotations for marking segments of data."""
+        return self._annotations
+
+    @annotations.setter
+    def annotations(self, annotations):
+        """Setter for annotations. Checks if they are inside the data range.
+
+        Parameters
+        ----------
+        annotations : Instance of mne.Annotations
+            Annotations to set.
+        """
+        if annotations is not None:
+            if not isinstance(annotations, Annotations):
+                raise ValueError('Annotations must be an instance of '
+                                 'mne.Annotations. Got %s.' % annotations)
+            meas_date = self.info['meas_date']
+            if meas_date is None:
+                meas_date = 0
+            elif not np.isscalar(meas_date):
+                if len(meas_date) > 1:
+                    meas_date = meas_date[0] + meas_date[1] / 1000000.
+            if annotations.orig_time is not None:
+                offset = (annotations.orig_time - meas_date -
+                          self.first_samp / self.info['sfreq'])
+            else:
+                offset = 0
+            omit_ind = list()
+            for ind, onset in enumerate(annotations.onset):
+                onset += offset
+                if onset > self.times[-1]:
+                    warn('Omitting annotation outside data range.')
+                    omit_ind.append(ind)
+                elif onset < self.times[0]:
+                    if onset + annotations.duration[ind] < self.times[0]:
+                        warn('Omitting annotation outside data range.')
+                        omit_ind.append(ind)
+                    else:
+                        warn('Annotation starting outside the data range. '
+                             'Limiting to the start of data.')
+                        duration = annotations.duration[ind] + onset
+                        annotations.duration[ind] = duration
+                        annotations.onset[ind] = self.times[0] - offset
+                elif onset + annotations.duration[ind] > self.times[-1]:
+                    warn('Annotation expanding outside the data range. '
+                         'Limiting to the end of data.')
+                    annotations.duration[ind] = self.times[-1] - onset
+            annotations.onset = np.delete(annotations.onset, omit_ind)
+            annotations.duration = np.delete(annotations.duration, omit_ind)
+            annotations.description = np.delete(annotations.description,
+                                                omit_ind)
+
+        self._annotations = annotations
 
     def __del__(self):
         # remove file for memmap
@@ -1332,6 +1388,10 @@ class _BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
             # slice and copy to avoid the reference to large array
             raw._data = raw._data[:, smin:smax + 1].copy()
         raw._update_times()
+        if raw.annotations is not None:
+            annotations = raw.annotations
+            annotations.onset -= tmin
+            raw.annotations = annotations
         return raw
 
     @verbose
@@ -1719,18 +1779,19 @@ class _BaseRaw(ProjMixin, ContainsMixin, UpdateChannelsMixin,
             self.preload = True
 
         # now combine information from each raw file to construct new self
+        annotations = self.annotations
         for r in raws:
             self._first_samps = np.r_[self._first_samps, r._first_samps]
             self._last_samps = np.r_[self._last_samps, r._last_samps]
             self._raw_extras += r._raw_extras
             self._filenames += r._filenames
-            self.annotations = _combine_annotations((self.annotations,
-                                                     r.annotations),
-                                                    self._last_samps,
-                                                    self._first_samps,
-                                                    self.info['sfreq'])
+            annotations = _combine_annotations((annotations, r.annotations),
+                                               self._last_samps,
+                                               self._first_samps,
+                                               self.info['sfreq'])
 
         self._update_times()
+        self.annotations = annotations
 
         if not (len(self._first_samps) == len(self._last_samps) ==
                 len(self._raw_extras) == len(self._filenames)):